+++ R/get_graph_server.R
@@ -11,6 +11,8 @@
 #'
 #' @return name of the file with the graph
 #' @md
+#'
+#' @importFrom rsvg rsvg_png rsvg_pdf rsvg_ps
 #' @export
 #'
 #' @examples
@@ -29,7 +31,7 @@
 get_graph_server <- function(
   x,
   file = NULL,
-  vector = TRUE,
+  force_png = FALSE,
   ...
 ){
 
@@ -42,41 +44,65 @@ get_graph_server <- function(
     stop("'x' has to be of class 'plantuml' or a 'character` vector containing the plantuml code!")
   }
 
-  if ( is.null(file) ) {
-    if (vector) {
-      file <- tempfile( fileext = ".svg")
-    } else {
-      file <- tempfile( fileext = ".png")
+  if (!is.null(file)) {
+    pos <- regexpr("\\.([[:alnum:]]+)$", file)
+    type <- ifelse( pos > -1L, substring(file, pos + 1L), "")
+    if (!(type %in% getPlantumlOption("supported_formats"))) {
+      stop(
+        "Type '", type, "' not supported through plantUML server!"
+      )
     }
+    if (force_png & type == "png") {
+      tmptype <- "png"
+    } else {
+      tmptype <- "svg"
+      force_png <- FALSE
+    }
+  } else {
+    tmptype <- ifelse(
+      force_png,
+      "png",
+      "svg"
+    )
+    type <- tmptype
   }
 
-  pos <- regexpr("\\.([[:alnum:]]+)$", file)
-  type <- ifelse( pos > -1L, substring(file, pos + 1L), "")
-
-  if (!(type %in% c("svg", "png", "txt"))) {
-    stop(
-      "Type ", type, " not supported buy plantUML server!",
-    )
-    result <- get_graph_local(
-      x = x,
-      file = file,
-      vector = vector,
-      ...
-    )
+  if (type == "txt") {
+    tmptype <- "txt"
   }
+
+  tmpfile <- tempfile( pattern = "plantuml.", fileext = paste0(".", tmptype))
 
   url <- plantuml_URL(
     plantuml = x,
-    type = type
+    type = tmptype
   )
 
-  result <- utils::download.file(url, file)
+  result <- utils::download.file(url, tmpfile)
+
   if (result != 0) {
     unlink(file)
     stop("Error in download!")
   } else {
-    result <- file
+    result <- tmpfile
   }
 
-  return(result)
+  if (force_png){
+    if (!is.null(file)) { file.copy( from = tmpfile, to = file, overwrite = TRUE) }
+  } else if (type != "svg") {
+    switch (
+      type,
+      png = rsvg::rsvg_png(svg = tmpfile, file = file),
+      pdf = rsvg::rsvg_pdf(svg = tmpfile, file = file),
+      ps  = rsvg::rsvg_ps(svg = tmpfile, file = file),
+      txt = if (!is.null(file)){ file.copy( from = tmpfile, to = file, overwrite = TRUE) },
+      stop("Not supported conversion!")
+    )
+  }
+
+  if (is.null(file)){
+    file <- tmpfile
+  }
+
+  return(file)
 }