+++ NEWS.md
@@ -3,6 +3,7 @@
 - included a function to generate the graph on a plantuml server, i.e. independant of a local installation. 
 - use Smetana instead of `dot` (see [here for details](https://plantuml.io/smetana02))
 - default vector format as `svg` instead of `eps`
+- Bump R dependency to >= 4.0
 
 # plantuml 0.3.5
 ## Major and Breaking changes